refactor(ale_claw): move computer_handler into harness/computer/ subpackage
Names the VM execution-substrate boundary (OpenClawComputerHandler +
MCPComputerHandler) instead of leaving it at base. mcp_runtime stays in
tools/ (it's a tool backend like fs_backends, imported 2:1 by tools).
Behavior-preserving: git mv + import-path updates (computer_handler's
.tools.mcp_runtime -> ..tools.mcp_runtime; importers -> .computer.*). 928/3.
